+++ FinalAssignment/Controllers/ScheduleController.cs
@@ -15,12 +15,18 @@ namespace FinalAssignment.Controllers
         private PracticeContext db = new PracticeContext();
 
         // GET: Schedule
+        [Authorize]
         public ActionResult Index()
         {
+            if (!(User.IsInRole("Admin") || User.IsInRole("Reception")))
+            {
+                return HttpNotFound();
+            }
             return View(db.Schedules.ToList());
         }
 
         // GET: Schedule/Details/5
+        [Authorize]
         public ActionResult Details(int? id)
         {
             if (id == null)
@@ -32,12 +38,21 @@ namespace FinalAssignment.Controllers
             {
                 return HttpNotFound();
             }
+            if (!(User.IsInRole("Admin") || User.IsInRole("Reception")))
+            {
+                return HttpNotFound();
+            }
             return View(schedule);
         }
 
         // GET: Schedule/Create
+        [Authorize]
         public ActionResult Create()
         {
+            if (!(User.IsInRole("Admin") || User.IsInRole("Reception")))
+            {
+                return HttpNotFound();
+            }
             return View();
         }
 
@@ -46,6 +61,7 @@ namespace FinalAssignment.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult Create([Bind(Include = "ScheduleID,Name,StartTime,EndTime")] Schedule schedule)
         {
             if (ModelState.IsValid)
@@ -59,6 +75,7 @@ namespace FinalAssignment.Controllers
         }
 
         // GET: Schedule/Edit/5
+        [Authorize]
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -70,12 +87,17 @@ namespace FinalAssignment.Controllers
             {
                 return HttpNotFound();
             }
+            if (!(User.IsInRole("Admin") || User.IsInRole("Reception")))
+            {
+                return HttpNotFound();
+            }
             return View(schedule);
         }
 
         // POST: Schedule/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ScheduleID,Name,StartTime,EndTime")] Schedule schedule)
@@ -90,6 +112,7 @@ namespace FinalAssignment.Controllers
         }
 
         // GET: Schedule/Delete/5
+        [Authorize]
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -101,14 +124,23 @@ namespace FinalAssignment.Controllers
             {
                 return HttpNotFound();
             }
+            if (!(User.IsInRole("Admin") || User.IsInRole("Reception")))
+            {
+                return HttpNotFound();
+            }
             return View(schedule);
         }
 
         // POST: Schedule/Delete/5
+        [Authorize]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!(User.IsInRole("Admin") || User.IsInRole("Reception")))
+            {
+                return HttpNotFound();
+            }
             Schedule schedule = db.Schedules.Find(id);
             db.Schedules.Remove(schedule);
             db.SaveChanges();
